+++ Contoso.Domain/Entities/VariableMetaDataModel.cs
@@ -1,0 +1,21 @@
+﻿using LogicBuilder.Attributes;
+using LogicBuilder.Domain;
+
+
+namespace Contoso.Domain.Entities
+{
+    public class VariableMetaDataModel : BaseModel
+    {
+		[VariableEditorControl(VariableControlType.SingleLineTextBox)]
+		[AlsoKnownAs("VariableMetaData_VariableMetaDataId")]
+		public int VariableMetaDataId { get; set; }
+
+		[VariableEditorControl(VariableControlType.SingleLineTextBox)]
+		[AlsoKnownAs("VariableMetaData_Data")]
+		public string Data { get; set; }
+
+		[VariableEditorControl(VariableControlType.SingleLineTextBox)]
+		[AlsoKnownAs("VariableMetaData_LastUpdated")]
+		public System.DateTime LastUpdated { get; set; }
+    }
+}