+++ Contoso.Test.Business/Requests/IBaseRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace Contoso.Test.Business.Requests
+{
+    public interface IBaseRequest
+    {
+    }
+}