+++ Contoso.Test.Business/Requests/SaveEntityRequest.cs
@@ -1,0 +1,9 @@
+﻿using LogicBuilder.Domain;
+
+namespace Contoso.Test.Business.Requests
+{
+    public class SaveEntityRequest : IBaseRequest
+    {
+        public BaseModel Entity { get; set; }
+    }
+}