+++ Contoso.Test.Business/Responses/Json/ResponseConverter.cs
@@ -1,0 +1,9 @@
+﻿using LogicBuilder.Domain.Json;
+
+namespace Contoso.Test.Business.Responses.Json
+{
+    public class ResponseConverter : JsonTypeConverter<BaseResponse>
+    {
+        public override string TypePropertyName => nameof(BaseResponse.TypeFullName);
+    }
+}