+++ Contoso.Test.Flow/IFlowManager.cs
@@ -1,0 +1,20 @@
+﻿using Contoso.Test.Flow.Cache;
+using LogicBuilder.RulesDirector;
+using System;
+
+namespace Contoso.Test.Flow
+{
+    public interface IFlowManager
+    {
+        DirectorBase Director { get; }
+        IFlowActivity FlowActivity { get; }
+        FlowDataCache FlowDataCache { get; }
+        Progress Progress { get; }
+        IServiceProvider ServiceProvider { get; }
+
+        void Start(string module);
+        void SetCurrentBusinessBackupData();
+        void FlowComplete();
+        void Terminate();
+    }
+}