+++ LogicBuilder.App.Utils.Tests/Rules/SaveDepartmentTest.cs
@@ -1,0 +1,112 @@
+﻿using Contoso.Domain.Entities;
+using Contoso.Test.Business.Requests;
+using Contoso.Test.Flow;
+using Contoso.Test.Flow.Cache;
+using LogicBuilder.RulesDirector;
+using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.Diagnostics.CodeAnalysis;
+
+namespace LogicBuilder.App.Utils.Tests.Rules
+{
+    public class SaveDepartmentTest
+    {
+        public SaveDepartmentTest(ITestOutputHelper output)
+        {
+            this.output = output;
+            Initialize();
+        }
+
+        #region Fields
+        private IServiceProvider serviceProvider;
+        private readonly ITestOutputHelper output;
+        #endregion Fields
+
+        [Fact]
+        public void SaveDepartment()
+        {
+            //arrange
+            IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
+            flowManager.FlowDataCache.Request = new SaveEntityRequest
+            {
+                Entity = new DepartmentModel
+                {
+                    EntityState = LogicBuilder.Domain.EntityStateType.Modified,
+                    InstructorID = 1,
+                    Budget = 10000,
+                    StartDate = new DateTime(2020, 2, 2, 0, 0, 0, DateTimeKind.Unspecified),
+                    Name = "Physics"
+                }
+            };
+
+            //act
+            System.Diagnostics.Stopwatch stopWatch = System.Diagnostics.Stopwatch.StartNew();
+            flowManager.Start("savedepartment");
+            stopWatch.Stop();
+            this.output.WriteLine("Saving valid department  = {0}", stopWatch.Elapsed.TotalMilliseconds);
+
+            //assert
+            Assert.True(flowManager.FlowDataCache.Response.Success);
+        }
+
+        [Fact]
+        public void SaveInvalidDepartment()
+        {
+            //arrange
+            IFlowManager flowManager = serviceProvider.GetRequiredService<IFlowManager>();
+            flowManager.FlowDataCache.Request = new SaveEntityRequest
+            {
+                Entity = new DepartmentModel
+                {
+                    EntityState = LogicBuilder.Domain.EntityStateType.Modified,
+                    InstructorID = null,
+                    Budget = -1,
+                    StartDate = new DateTime(1, 1, 1, 0, 0, 0, DateTimeKind.Unspecified),
+                    Name = ""
+                }
+            };
+
+            //act
+            System.Diagnostics.Stopwatch stopWatch = System.Diagnostics.Stopwatch.StartNew();
+            flowManager.Start("savedepartment");
+            stopWatch.Stop();
+            this.output.WriteLine("Saving invalid department  = {0}", stopWatch.Elapsed.TotalMilliseconds);
+
+            //assert
+            Assert.False(flowManager.FlowDataCache.Response.Success);
+            Assert.Equal(4, flowManager.FlowDataCache.Response.ErrorMessages.Count);
+        }
+
+        #region Helpers
+        [MemberNotNull(nameof(serviceProvider))]
+        private void Initialize()
+        {
+            serviceProvider = new ServiceCollection()
+                .AddLogging()
+                .AddTransient<IFlowManager, FlowManager>()
+                .AddTransient<DirectorFactory, DirectorFactory>()
+                .AddTransient<ICustomActions, CustomActions>()
+                .AddScoped<FlowDataCache>()
+                .AddScoped<Progress>()
+                .AddAppUtilsGenericsHelpers()
+                .AddAppUtilsObjectHelper()
+                .AddRulesCacheService
+                (
+                    new Utils.Rules.RulesLoaderRequest
+                    (
+                        "Contoso.Test.Flow.Rulesets",
+                        typeof(FlowActivity),
+                        [
+                            typeof(Interfaces.ITypeHelper).Assembly,
+                            typeof(IBaseRequest).Assembly,
+                            typeof(CourseModel).Assembly,
+                            typeof(DirectorBase).Assembly,
+                            typeof(string).Assembly
+                        ]
+                    )
+                )
+                .BuildServiceProvider();
+        }
+        #endregion Helpers
+    }
+}