+++ LogicBuilder.App.Utils/Rules/Interfaces/IRulesSerializer.cs
@@ -1,0 +1,11 @@
+﻿using LogicBuilder.Workflow.Activities.Rules;
+
+namespace LogicBuilder.App.Utils.Rules.Interfaces
+{
+    public interface IRulesSerializer
+    {
+        RuleSet? DeserializeRuleSet(string ruleSetXmlDefinition);
+        RuleSet? DeserializeRuleSetFile(RulesModule module);
+        RuleValidation GetValidation(RuleSet ruleSet, RulesLoaderRequest rulesLoaderRequest);
+    }
+}